+++ ProgramsCore/ProgramInstance.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace DiscordGameEngine.Core
+{
+    public class ProgramInstance
+    {
+
+        public Discord.WebSocket.SocketUser author;
+
+    }
+}