+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-﻿#V0.1.0 :
+﻿#V0.2.2
+Created `LogManager` : contains LogPrefixes (WARN, LOG, ERROR, DEBUG) and has the `LogDebug` method moved from Program
+Created `ProgramsInstance` and `ProgramsManager` which will Serve as holders for programs made with the engine
+Added the Method `GetArgsAsSingleStringFrom` read the documention in the XML file after building/code summary for more information
+
+#V0.1.0 :
 Extended base + command support : ready to implement features
 
 #V0.0.0 :